+++ com.unity.mobile.notifications/Runtime/Android/Plugins/com/unity/androidnotifications/UnityNotificationManager.java
@@ -31,7 +31,6 @@ import static android.app.Notification.VISIBILITY_PUBLIC;
 
 import java.io.InputStream;
 import java.lang.Integer;
-import java.lang.reflect.Method;
 import java.util.Calendar;
 import java.util.Random;
 import java.util.Set;
@@ -55,10 +54,6 @@ public class UnityNotificationManager extends BroadcastReceiver {
     private ConcurrentHashMap<Integer, Notification.Builder> mScheduledNotifications;
     private NotificationCallback mNotificationCallback;
     private int mExactSchedulingSetting = -1;
-    private Method mCanScheduleExactAlarms;
-    private Method mBigIcon;
-    private Method mSetContentDescription;
-    private Method mShowBigPictureWhenCollapsed;
 
     static final String TAG_UNITY = "UnityNotifications";
 
@@ -504,14 +499,14 @@ public class UnityNotificationManager extends BroadcastReceiver {
     }
 
     private PendingIntent getActivityPendingIntent(int id, Intent intent, int flags) {
-        if (Build.VERSION.SDK_INT >= 23)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M)
             return PendingIntent.getActivity(mContext, id, intent, flags | PendingIntent.FLAG_IMMUTABLE);
         else
             return PendingIntent.getActivity(mContext, id, intent, flags);
     }
 
     private PendingIntent getBroadcastPendingIntent(int id, Intent intent, int flags) {
-        if (Build.VERSION.SDK_INT >= 23)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M)
             return PendingIntent.getBroadcast(mContext, id, intent, flags | PendingIntent.FLAG_IMMUTABLE);
         else
             return PendingIntent.getBroadcast(mContext, id, intent, flags);
@@ -579,17 +574,7 @@ public class UnityNotificationManager extends BroadcastReceiver {
         if (Build.VERSION.SDK_INT < 31)
             return true;
 
-        try {
-            if (mCanScheduleExactAlarms == null)
-                mCanScheduleExactAlarms = AlarmManager.class.getMethod("canScheduleExactAlarms");
-            return (boolean)mCanScheduleExactAlarms.invoke(alarmManager);
-        } catch (NoSuchMethodException ex) {
-            Log.e(TAG_UNITY, "No AlarmManager.canScheduleExactAlarms() on Android 31+ device, should not happen", ex);
-            return false;
-        } catch (Exception ex) {
-            Log.e(TAG_UNITY, "AlarmManager.canScheduleExactAlarms() threw", ex);
-            return false;
-        }
+        return alarmManager.canScheduleExactAlarms();
     }
 
     public boolean canScheduleExactAlarms() {
@@ -759,7 +744,8 @@ public class UnityNotificationManager extends BroadcastReceiver {
         }
 
         try {
-            mNotificationCallback.onSentNotification(notification);
+            if (mNotificationCallback != null)
+                mNotificationCallback.onSentNotification(notification);
         } catch (RuntimeException ex) {
             Log.w(TAG_UNITY, "Can not invoke OnNotificationReceived event when the app is not running!");
         }
@@ -936,9 +922,9 @@ public class UnityNotificationManager extends BroadcastReceiver {
         if (picture.charAt(0) == '/') {
             style.bigPicture(BitmapFactory.decodeFile(picture));
         } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && picture.indexOf("://") > 0) {
-            if (Build.VERSION.SDK_INT >= 31) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
                 Icon icon = Icon.createWithContentUri(picture);
-                setBigPictureIcon(style, icon);
+                style.bigPicture(icon);
             } else {
                 Bitmap pic = loadBitmap(picture);
                 if (pic != null) {
@@ -946,57 +932,21 @@ public class UnityNotificationManager extends BroadcastReceiver {
                 }
             }
         } else {
-            Object pic = getIconFromResources(picture, Build.VERSION.SDK_INT < 31);
-            if (Build.VERSION.SDK_INT >= 31 && pic instanceof Icon)
-                setBigPictureIcon(style, pic);
+            Object pic = getIconFromResources(picture, Build.VERSION.SDK_INT < Build.VERSION_CODES.S);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S && pic instanceof Icon)
+                style.bigPicture((Icon)pic);
             else if (pic instanceof Bitmap)
                 style.bigPicture((Bitmap)pic);
         }
 
         style.setBigContentTitle(extras.getString(KEY_BIG_CONTENT_TITLE));
         style.setSummaryText(extras.getString(KEY_BIG_SUMMARY_TEXT));
-        if (Build.VERSION.SDK_INT >= 31) {
-            setBigPictureProps31(style, extras.getString(KEY_BIG_CONTENT_DESCRIPTION), extras.getBoolean(KEY_BIG_SHOW_WHEN_COLLAPSED, false));
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+            style.setContentDescription(extras.getString(KEY_BIG_CONTENT_DESCRIPTION));
+            style.showBigPictureWhenCollapsed(extras.getBoolean(KEY_BIG_SHOW_WHEN_COLLAPSED, false));
         }
 
         builder.setStyle(style);
-    }
-
-    @TargetApi(Build.VERSION_CODES.M)
-    private void setBigPictureIcon(Notification.BigPictureStyle style, Object icon) {
-        if (mBigIcon == null) {
-            try {
-                mBigIcon = Notification.BigPictureStyle.class.getMethod("bigPicture", Icon.class);
-            } catch (NoSuchMethodException e) {
-                Log.e(TAG_UNITY, "Failed to find method bigPicture", e);
-                return;
-            }
-        }
-
-        try {
-            mBigIcon.invoke(style, icon);
-        } catch (Exception e) {
-            Log.e(TAG_UNITY, "Failed to call method bigPicture", e);
-        }
-    }
-
-    private void setBigPictureProps31(Notification.BigPictureStyle style, String contentDesc, boolean showWhenCollapsed) {
-        if (mSetContentDescription == null || mShowBigPictureWhenCollapsed == null) {
-            try {
-                mSetContentDescription = Notification.BigPictureStyle.class.getMethod("setContentDescription", CharSequence.class);
-                mShowBigPictureWhenCollapsed = Notification.BigPictureStyle.class.getMethod("showBigPictureWhenCollapsed", boolean.class);
-            } catch (NoSuchMethodException e) {
-                Log.e(TAG_UNITY, "Failed to find method", e);
-                return;
-            }
-        }
-
-        try {
-            mSetContentDescription.invoke(style, contentDesc);
-            mShowBigPictureWhenCollapsed.invoke(style, showWhenCollapsed);
-        } catch (Exception e) {
-            Log.e(TAG_UNITY, "Failed to call method", e);
-        }
     }
 
     public static void setNotificationColor(Notification.Builder notificationBuilder, int color) {